Added function for generating synthetic data starting from a pdf
The function for generating data isn't fully correct. The generation of two univariate correlated series isn't correct beacuse the two series aren't generated from the same distribution. Furthermore the generation of the bivariate series has to be check/edit before using just to be sure that the data generated are taken from the wanted bivariate distribution.